requirements.txt
google-adk>=0.0.1
google-generativeai>=0.3.0
tavily-python>=0.3.0
gradio>=4.0.0
python-dotenv>=1.0.0